upd group many things compiled!!!!!

+++ src/main/java/tw/idv/cha102/g7/group/service/Impl/GroupServiceImpl.java
@@ -9,12 +9,17 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Component;
 import tw.idv.cha102.g7.group.dto.*;
 import tw.idv.cha102.g7.group.entity.Group;
+import tw.idv.cha102.g7.group.entity.MemberDetail;
 import tw.idv.cha102.g7.group.repo.GroupRepository;
+import tw.idv.cha102.g7.group.repo.MemberDetailRepository;
+import tw.idv.cha102.g7.group.repo.RegFormRepository;
 import tw.idv.cha102.g7.group.service.GroupService;
 
 import java.math.BigInteger;
+import java.sql.Timestamp;
 import java.util.Date;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Stream;
 
 @Component
@@ -22,6 +27,12 @@ public class GroupServiceImpl implements GroupService {
 
     @Autowired
     private GroupRepository groupRepository;
+
+    @Autowired
+    private RegFormRepository regFormRepository;
+
+    @Autowired
+    private MemberDetailRepository memberDetailRepository;
 
     public void insert(Group group) {
         groupRepository.save(group);
@@ -53,7 +64,21 @@ public class GroupServiceImpl implements GroupService {
     }
 
     public void delete(Integer groupId) {
-        groupRepository.deleteById(groupId);
+        boolean control = true;
+        List<RegFormMemberDetailDto> refundList = regFormRepository.findRefundSTaByGroupId(groupId);
+
+        //如果還有尚未退費的，就不能刪除
+        if (refundList != null) {
+            for (RegFormMemberDetailDto refundLists : refundList) {
+                if (refundLists.getRefund_Sta() != 3) {
+                    control = false;
+                }
+            }
+        }
+
+        if (control) {
+            groupRepository.deleteById(groupId);
+        }
     }
 
     public Group getGroupByGroupId(Integer groupId) {
@@ -99,8 +124,10 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public List<GroupRegFormDto> findGroupRegFormDtoByGroupId(Integer groupId) {
-        return groupRepository.findGroupRegFormDtoByGroupId(groupId);
+    public Stream<GroupRegFormDto> findGroupRegFormDtoByMemId(Integer memId, Integer page) {
+        Sort sort = Sort.by(Sort.Direction.ASC, "Group_Sta");
+        Pageable pageable = PageRequest.of(page, 3, sort);
+        return groupRepository.findGroupRegFormDtoByMemId(memId, pageable).get();
     }
 
     @Override
@@ -109,7 +136,7 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public Stream<GroupListDto> findGroupListByGroupSta(Integer groupSta, Integer page) {
+    public Stream<GroupListDto> findGroupListByGroupSta(Integer page) {
 
         Sort sort = Sort.by(Sort.Direction.ASC, "start_Date");  //定義Sort ASC為升冪排序，通過揪團狀態排序 start_Date 是MySQL的欄位名
 
@@ -129,14 +156,15 @@ public class GroupServiceImpl implements GroupService {
 
         Pageable pageable = PageRequest.of(page, 6, sort);   //定義Pageable(page 當前頁數, size 傳入資料筆數, sort 排序方法)
         BigInteger zero = new BigInteger("0");  //比較BigInteger所以要用BigInteger互相比較
-
+        Long datetime = System.currentTimeMillis();
+        Timestamp timestamp = new Timestamp(datetime);
         //以下判斷取出的揪團狀態
         List<Group> groups = groupRepository.findGroupByGroupSta(0); //找出揪團狀態為0的列表
         for (Group group : groups) {
 
             //參團人數到最大人數
             if (group.getMembers() == group.getMaxMember()) {
-                group.setGroupSta(1); //設置成團狀態
+                group.setGroupSta(5); //設置額滿狀態
                 groupRepository.save(group);
             }
 
@@ -149,17 +177,49 @@ public class GroupServiceImpl implements GroupService {
                 } else {
                     group.setGroupSta(2); //設置揪團取消狀態
                     groupRepository.save(group);
+
+                    //將參團人員設置申請退款
+                    List<GroupRegFormMemberDetailDto> detailList = memberDetailRepository.findGroupRegFormMemberDetailDtoByGroupIdOrderByFormId(group.getGroupId());
+                    if (detailList != null) {
+                        for (GroupRegFormMemberDetailDto detail : detailList) {
+                            MemberDetail updDetail = memberDetailRepository.findById(detail.getDetailId()).get();
+                            if (updDetail != null) {
+                                if (updDetail.getRefund() == 0 || updDetail.getRefund() == 4) {
+                                    updDetail.setRefundSta(2);
+                                    updDetail.setRefund(group.getAmount());
+                                    updDetail.setRefundDate(timestamp);
+                                    memberDetailRepository.save(updDetail);
+                                }
+                            }
+                        }
+                    }
+
                 }
             }
         }
         //問題 Stream.filter()後的列表是亂的(結果:因為pageable size一次傳入固定比數， 呼叫分頁時才會再給下一頁的值，所以才會挖空)
-        return groupRepository.findGroupListByGroupSta(groupSta, pageable).get(); //返回值為Stream<T>
+        return groupRepository.findGroupListByGroupSta(pageable).get(); //返回值為Stream<T>
     }
 
     @Override
     public Stream<MyGroupListDto> findMyGroupListDtoByMemId(Integer memId, Integer page) {
         Sort sort = Sort.by(Sort.Direction.ASC, "Group_Sta");
         Pageable pageable = PageRequest.of(page, 6, sort);
+
+        List<Group> groupList = groupRepository.findByMemIdOrderByGroupStaDesc(memId);
+        for (Group group : groupList) {
+            //揪團時間截止做的判斷
+            Date date = group.getDeadline(); //轉換java.sql.TimeStamp 至 java.util.Date
+            if (date.compareTo(new Date()) < 0) {
+                if (group.getMembers() > group.getMinMember()) {
+                    group.setGroupSta(1); //設置成團狀態
+                    groupRepository.save(group);
+                } else {
+                    group.setGroupSta(2); //設置揪團取消狀態
+                    groupRepository.save(group);
+                }
+            }
+        }
         return groupRepository.findMyGroupListDtoByMemId(memId, pageable).get();
     }
 
@@ -169,43 +229,49 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public Stream<GroupListDto> findGroupByGroupStaOrderByDeadline(Integer groupSta, Integer page) {
+    public Stream<GroupListDto> findGroupByGroupStaOrderByDeadline(Integer page) {
         Pageable pageable = PageRequest.of(page, 6);
-        return groupRepository.findGroupByGroupStaOrderByDeadline(groupSta, pageable).get();
+        return groupRepository.findGroupByGroupStaOrderByDeadline(pageable).get();
     }
 
     @Override
-    public Stream<GroupListDto> findGroupByGroupStaOrderByDeadlineDesc(Integer groupSta, Integer page) {
+    public Stream<GroupListDto> findGroupByGroupStaOrderByDeadlineDesc(Integer page) {
         Pageable pageable = PageRequest.of(page, 6);
-        return groupRepository.findGroupByGroupStaOrderByDeadlineDesc(groupSta, pageable).get();
+        return groupRepository.findGroupByGroupStaOrderByDeadlineDesc(pageable).get();
     }
 
     @Override
-    public Stream<GroupListDto> findGroupByGroupStaOrderByAmount(Integer groupSta, Integer page) {
+    public Stream<GroupListDto> findGroupByGroupStaOrderByAmount(Integer page) {
         Pageable pageable = PageRequest.of(page, 6);
-        return groupRepository.findGroupByGroupStaOrderByAmount(groupSta, pageable).get();
+        return groupRepository.findGroupByGroupStaOrderByAmount(pageable).get();
     }
 
     @Override
-    public Stream<GroupListDto> findGroupByGroupStaOrderByAmountDesc(Integer groupSta, Integer page) {
+    public Stream<GroupListDto> findGroupByGroupStaOrderByAmountDesc(Integer page) {
         Pageable pageable = PageRequest.of(page, 6);
-        return groupRepository.findGroupByGroupStaOrderByAmountDesc(groupSta, pageable).get();
+        return groupRepository.findGroupByGroupStaOrderByAmountDesc(pageable).get();
     }
 
     @Override
-    public Stream<GroupListDto> findGroupByGroupStaThemeLike(Integer groupSta, String str, Integer page) {
+    public Stream<GroupListDto> findGroupByGroupStaThemeLike(String str, Integer page) {
         Sort sort = Sort.by(Sort.Direction.ASC, "start_Date");
         Pageable pageable = PageRequest.of(page, 6, sort);
-        return groupRepository.findGroupByGroupStaThemeLike(groupSta, str, pageable).get();
+        return groupRepository.findGroupByGroupStaThemeLike(str, pageable).get();
     }
 
     @Override
     public GroupMemo findGroupMemo(Integer groupId) {
         Group group = groupRepository.findById(groupId).get();
 
-        //如果人數達標則自動設置成團狀態
-        if(group.getMembers() == group.getMaxMember()){
+        //設置人數達標
+        if (group.getMembers() >= group.getMinMember() && group.getMembers() < group.getMaxMember()) {
             group.setGroupSta(1);
+            groupRepository.save(group);
+        }
+
+        //如果人數達標則自動設置成團狀態
+        if (group.getMembers() == group.getMaxMember()) {
+            group.setGroupSta(5);
             groupRepository.save(group);
         }
 
@@ -217,5 +283,97 @@ public class GroupServiceImpl implements GroupService {
         return groupRepository.finGroupMember(groupId);
     }
 
+    @Override
+    public Stream<GroupRegFormDto> findGroupRegFormDtoByMemIdAndGroupSta0(Integer memId, Integer page) {
+        Sort sort = Sort.by(Sort.Direction.ASC, "Dep_Date");
+        Pageable pageable = PageRequest.of(page, 3, sort);
+        return groupRepository.findGroupRegFormDtoByMemIdAndGroupSta0(memId, pageable).get();
+    }
+
+    @Override
+    public Stream<GroupRegFormDto> findGroupRegFormDtoByMemIdAndGroupSta1(Integer memId, Integer page) {
+        Sort sort = Sort.by(Sort.Direction.ASC, "Dep_Date");
+        Pageable pageable = PageRequest.of(page, 3, sort);
+        return groupRepository.findGroupRegFormDtoByMemIdAndGroupSta1(memId, pageable).get();
+    }
+
+    @Override
+    public Stream<GroupRegFormDto> findGroupRegFormDtoByMemIdAndThemeLike(Integer memId, String str, Integer page) {
+        Sort sort = Sort.by(Sort.Direction.ASC, "Group_Sta");
+        Pageable pageable = PageRequest.of(page, 3, sort);
+        return groupRepository.findGroupRegFormDtoByMemIdAndThemeLike(memId, str, pageable).get();
+    }
+
+    @Override
+    public void updateGroupSta(Integer groupId, Group group) {
+        var updGroup = groupRepository.findById(groupId);
+        if (updGroup.isPresent()) {
+            updGroup.get().setGroupSta(group.getGroupSta());
+            groupRepository.save(updGroup.get());
+        }
+    }
+
+    @Override
+    public void updateGroupStaPaymentSta(Integer groupId, Group group) {
+        var updGroup = groupRepository.findById(groupId);
+        Long datetime = System.currentTimeMillis();
+        Timestamp timestamp = new Timestamp(datetime);
+
+        if (updGroup.isPresent()) {
+            if (group.getGroupSta() != null) {
+                updGroup.get().setGroupSta(group.getGroupSta());
+            }
+            if (group.getPaymentSta() != null) {
+                updGroup.get().setPaymentSta(group.getPaymentSta());
+            }
+            groupRepository.save(updGroup.get());
+        }
+
+        //如果管理員設置取消揪團或下架揪團時
+        if (group.getGroupSta() == 2 || group.getGroupSta() == 4) {
+            List<MemberDetail> memberDetailList = memberDetailRepository.findByGroupId(groupId);
+            if (memberDetailList != null) {
+                //設置退費狀態
+                for (MemberDetail detail : memberDetailList) {
+                    if (detail.getRefundSta() == 0) {
+                        detail.setRefundSta(2);
+                        detail.setRefundDate(timestamp);
+                        detail.setRefund(group.getAmount());
+                    }
+                }
+            }
+        }
+    }
+
+    //    為了多表查members只在service用
+    @Override
+    public Group findMemberDeadlineAmountByDetailId(Integer detailId) {
+        return groupRepository.findMemberDeadlineAmountByDetailId(detailId);
+    }
+
+    @Override
+    public List<Group> findAll() {
+        List<Group> groupList = groupRepository.findAll();
+        TimeUnit time = TimeUnit.DAYS;
+        for (Group group : groupList) {
+            Date depDate = group.getDepDate();
+            long diff = new Date().getTime() - depDate.getTime();
+            long difference = time.convert(diff, TimeUnit.MILLISECONDS);
+            if(difference >= 14){
+                group.setPaymentSta(1);
+            }
+
+            List<MemberDetail> memberDetailList = memberDetailRepository.findByGroupId(group.getGroupId());
+            for (MemberDetail detail : memberDetailList) {
+                if (detail != null) {
+                    if (detail.getRefundSta() == 2) { //如果還有申請退款的團員
+                        group.setPaymentSta(2); //設置不能請款
+                    }
+                }
+            }
+        }
+
+        return groupList;
+    }
 
 }
